// src/app/models/product.model.ts
export class Product {
	_id?: any;
	id?: any;
	sku?: string;
	name?: string;
	description?: string;
	barcode?: string;
	suom?: any;
	puom?: any;
	fg?: boolean;
	rm?: boolean;
	image?: string;
    listprice?: number;
    botprice?: number;
    cost?: number;
    min?: number;
    max?: number;
    supplier?: any;
    isStock?: boolean;
	category?: any;
	taxin?: any;
	taxout?: any;
	brand?: any;
	active?: boolean;
}